Move QuickButtonPrefs to not be in dedicated package
