+++ app/src/main/java/com/jkuester/unlauncher/datasource/DataStoreModule.kt
@@ -3,9 +3,7 @@ package com.jkuester.unlauncher.datasource
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.dataStore
-import com.jkuester.unlauncher.datasource.quickbuttonprefs.QuickButtonPreferencesSerializer
-import com.jkuester.unlauncher.datasource.quickbuttonprefs.ToThreeQuickButtonsMigration
-import com.jkuester.unlauncher.datasource.quickbuttonprefs.sharedPrefsMigration as quickButtonSharedPrefsMigration
+import com.jkuester.unlauncher.datasource.sharedPrefsMigration as quickButtonSharedPrefsMigration
 import com.jkuester.unlauncher.datastore.proto.CorePreferences
 import com.jkuester.unlauncher.datastore.proto.QuickButtonPreferences
 import com.jkuester.unlauncher.datastore.proto.UnlauncherApps
